+++ Mods ChoGGi/Solylent Glutton Provides Double/metadata.lua
@@ -1,20 +1,21 @@
 return PlaceObj("ModDef", {
 	"title", "Solylent Glutton Provides Double",
-	"version", 1,
-	"version_major", 0,
-	"version_minor", 1,
-
-	"image", "Preview.png",
 	"id", "ChoGGi_SolylentGluttonProvidesDouble",
 	"steam_id", "1907034997",
 	"pops_any_uuid", "32c54f4c-6008-4644-882c-5cc8403a08a9",
+	"version", 2,
+	"version_major", 0,
+	"version_minor", 2,
+	"image", "Preview.png",
 	"author", "ChoGGi",
 	"lua_revision", 1007000, -- Picard
 	"code", {
 		"Code/Script.lua",
 	},
-	"description", [[If you have the solylent green breakthrough; gluttons will provide double the soylent.
+	"description", [[
+If you have the solylent green breakthrough; gluttons will provide double the soylent.
 
 
-Not exactly requested by Monovigma...]],
+Not exactly requested by Monovigma...
+]],
 })